feat(SMTP): Add new `SMTP_SENDER` parameter (refs #7014)
On installations where mails could be emitted with arbitrary "From:"
users' addresses (user's address not belonging to the same domain as the
Dynacase's server), messages could be rejected by the recipients' MXs
due to SPF rules prohibiting messages from being sent with that domain
from the non-authorized Dynacase's server.

To circumvent these restrictions, a new SMTP_SENDER parameter as been
added to specify a fixed address (e.g. 'administrator@example.net') that
will be used for "MAIL FROM:" and "From:" when sending the message,
whereas the original user's address will appears in a "Reply-To:"
header.