fix(getLatestTDoc): Use the newest `fromid` (refs #7274)
`getLatestTDoc` would not work correctly when working on converted
document (i.e. when a document was converted from one family to
another).

The problem was that the `fromid` obtained from the document's `initid`
was non-deterministic and hence could be the `fromid` from the old
family instead of the `fromid` from the current family.