fix(VaultDiskFs): Correct computation of vault's dir size in findFreeFS (refs #7283)
The computation of vault's dir size (in findFreeFS) was erroneous when a
dir existed with no files in it (these case occurs when only temporary
files are created in the dir and the cleanContext script removes the
temp files from that dir).

So, the query has been changed to handle this case by using a "RIGHT
JOIN" to return the dir entries even when they contains no files (with
sum being 0 instead of NULL).

Also:
- Correct a possible mistake regarding computation in the case of
  multiples vaults (the JOIN must be done on both `id_dir` AND `id_fs`,
  and not on `id_dir` alone).